feat: 예외 처리 추가
InvalidInputException 추가하여 null값이 있는 경우에 대하여 오류 메시지 전송, InvalidRoutineExceiption 추가하여 권한이 있는 유저만 루틴 수정/삭제 가능하게 함.

issue #5

+++ IrumOrder_BackEnd/src/main/java/Irumping/IrumOrder/service/RoutineService.java
@@ -2,6 +2,8 @@ package Irumping.IrumOrder.service;
 
 import Irumping.IrumOrder.dto.RoutineDto;
 import Irumping.IrumOrder.entity.RoutineEntity;
+import Irumping.IrumOrder.exeption.InvalidInputException;
+import Irumping.IrumOrder.exeption.InvalidRoutineExceiption;
 import Irumping.IrumOrder.exeption.UserIdMismatchException;
 import Irumping.IrumOrder.repository.RoutineRepository;
 import jakarta.transaction.Transactional;
@@ -21,17 +23,32 @@ public class RoutineService {
         this.routineRepository = routineRepository;
     }
 
-    public List<RoutineEntity> getRoutinesByUserId(int userId) {
+    public List<RoutineEntity> getRoutinesByUserId(long userId) {
         return routineRepository.findByUserId(userId);
     }
 
     @Transactional
-    public RoutineEntity addRoutine(RoutineDto routineDto, int userId) {
+    public RoutineEntity addRoutine(RoutineDto routineDto, long userId) {
         if (userId!=(routineDto.getUserId())) {
             throw new UserIdMismatchException("User ID in request does not match the authenticated user ID.");
         }
         RoutineEntity routine = new RoutineEntity();
         routine.setUserId(userId);
+        if(routineDto.getMenuDetailId() == null){
+            throw new InvalidInputException("Menu detail id is required.");
+        }
+        if(routineDto.getMenuId() == null){
+            throw new InvalidInputException("Menu id is required.");
+        }
+        if(routineDto.getRoutineDay() == null){
+            throw new InvalidInputException("Routine Day  is required.");
+        }
+        if(routineDto.getRoutineTime() == null){
+            throw new InvalidInputException("Routine Time is required.");
+        }
+        if(routineDto.getIsActivated() == null){
+            throw new InvalidInputException("Alarm on/off value is required.");
+        }
         routine.setMenuId(routineDto.getMenuId());
         routine.setMenuDetailId(routineDto.getMenuDetailId());
         routine.setRoutineDay(routineDto.getRoutineDay());
@@ -44,8 +61,10 @@ public class RoutineService {
     @Transactional
     public RoutineEntity updateRoutine(Integer routineId, RoutineDto routineDto) {
         RoutineEntity routine = routineRepository.findById(routineId)
-                .orElseThrow(() -> new IllegalArgumentException("Routine not found with ID: " + routineId));
-
+                .orElseThrow(() -> new InvalidRoutineExceiption("Routine not found with ID: " + routineId));
+        if(routine.getUserId() != routineDto.getUserId()){
+            throw new UserIdMismatchException("User ID in request does not match the authenticated user ID.");
+        }
         if (routineDto.getMenuId() != null) {
             routine.setMenuId(routineDto.getMenuId());
         }
@@ -71,11 +90,13 @@ public class RoutineService {
 
 
     @Transactional
-    public void deleteRoutine(Integer routineId) {
+    public void deleteRoutine(Integer routineId, long userId) {
         // 루틴 존재 여부 확인 후 삭제
         RoutineEntity routine = routineRepository.findById(routineId)
-                .orElseThrow(() -> new IllegalArgumentException("Routine not found with ID: " + routineId));
-
+                .orElseThrow(() -> new InvalidRoutineExceiption("Routine not found with ID: " + routineId));
+        if(routine.getUserId() != userId){
+            throw new UserIdMismatchException("User ID in request does not match the authenticated user ID.");
+        }
         routineRepository.delete(routine);
     }
 }
